weShip MERN - Authentication - created requiredAuth middleware, protected routes, and ensure only users can create, update and delete deliveries

+++ backend/src/controllers/deliveryControllers.ts
@@ -3,10 +3,15 @@ import  asyncHandler from 'express-async-handler';
 import mongoose from 'mongoose';
 
 import Delivery from '../models/Delivery';
+import User from '../models/User';
+import { User as UserType } from '../types';
+
+
 
 // get all deliveries
 export const getDeliveryList = asyncHandler(async (req: Request, res: Response) => {
-  const deliveries = await Delivery.find();
+  const { user_id } = req.body;
+  const deliveries = await Delivery.find({ user_id });
   res.status(200).json({ deliveries });
 });
 
@@ -18,17 +23,26 @@ export const getDelivery = async (req: Request, res: Response) => {
     return res.status(404).json({ error: 'Invalid id, no such delivery record' });
   }
 
+  const user: UserType | null = await User.findById(req.body.user_id);
+  if (!user) {
+    return res.status(401).json({ error: 'User not found' });
+  }
+
   const delivery = await Delivery.findById(id);
   if (!delivery) {
     return res.status(404).json({ error: 'No delivery record found' });
   }
 
-  res.status(200).json({ delivery });
+  if (String(delivery.user_id) !== String(user._id)) {
+    return res.status(401).json({ error: 'User not authorized' });
+  }
+
+  return res.status(200).json({ delivery });
 };
 
 // create a delivery
 export const createDelivery = async (req: Request, res: Response) => {
-  const { customerName, deliveryDate, warehouseAddress, deliveryAddress } = req.body;
+  const { customerName, warehouseAddress, deliveryDate, deliveryAddress } = req.body;
   const fields = [];
 
   if (!customerName) fields.push(customerName);
@@ -42,10 +56,10 @@ export const createDelivery = async (req: Request, res: Response) => {
   
   try {
     const delivery = await Delivery.create(req.body);
-    res.status(200).json({ delivery });
+    return res.status(200).json({ delivery });
   } catch (error) {
     if (error instanceof Error) {
-      res.status(500).json({ error: error.message})
+      return res.status(500).json({ error: error.message})
     }
   }
 };
@@ -57,13 +71,30 @@ export const deleteDelivery = async (req: Request, res: Response) => {
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return res.status(404).json({ error: 'Invalid id, no such delivery record' });
   }
-  
-  const delivery = await Delivery.findByIdAndDelete(id);
+
+  const user: UserType | null = await User.findById(req.body.user_id);
+  if (!user) {
+    return res.status(401).json({ error: 'User not found' });
+  }
+
+  const delivery = await Delivery.findById(id);
   if (!delivery) {
     return res.status(404).json({ error: 'No delivery record found' });
   }
 
-  res.status(200).json({ delivery });
+  if (String(delivery.user_id) !== String(user._id)) {
+    return res.status(401).json({ error: 'User not authorized' });
+  }
+
+  try {
+    const deletedDelivery = await Delivery.findByIdAndDelete(id);
+    return res.status(200).json({ delivery: deletedDelivery });
+  } catch (error) {
+    console.error(error);
+    if (error instanceof Error) {
+      return res.status(500).json({ error: error.message})
+    }
+  }
 };
 
 // update a delivery
@@ -74,10 +105,27 @@ export const updateDelivery = async (req: Request, res: Response) => {
     return res.status(404).json({ error: 'Invalid id, no such delivery record' });
   }
 
-  const delivery = await Delivery.findByIdAndUpdate(id, req.body, { new: true });
+  const user: UserType | null = await User.findById(req.body.user_id);
+  if (!user) {
+    return res.status(401).json({ error: 'User not found' });
+  }
+
+  const delivery = await Delivery.findById(id);
   if (!delivery) {
     return res.status(404).json({ error: 'No delivery record found' });
   }
 
-  res.status(200).json({ delivery });
+  if (String(delivery.user_id) !== String(user._id)) {
+    return res.status(401).json({ error: 'User not authorized' });
+  }
+
+  try {
+    const updatedDelivery = await Delivery.findByIdAndUpdate(id, req.body, { new: true });
+    return res.status(200).json({ delivery: updatedDelivery });
+  } catch (error) {
+    console.error(error);
+    if (error instanceof Error) {
+      return res.status(500).json({ error: error.message});
+    }
+  }
 };
